requirements.txt
aiogram==2.13
python-dotenv